Se integra el plugin swaggger para una mejor documentación

+++ back/fisibet-apuestas/src/main/java/com/microservicio/fisibet/infraestructure/port/spring/TicketSpringPort.java
@@ -4,7 +4,7 @@ import com.microservicio.fisibet.infraestructure.model.TicketModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface TicketSpringPort extends JpaRepository<TicketModel, Long> {
+public interface TicketSpringPort extends JpaRepository<TicketModel, Integer> {
     @Query(value = "SELECT COUNT(*) FROM ticket", nativeQuery = true)
     Integer getCountRows();
 }
