+++ back/fisibet-usuario/src/main/java/com/microservicio/fisibet/infraestructure/port/spring/AccountUserSpringPort.java
@@ -3,5 +3,5 @@ package com.microservicio.fisibet.infraestructure.port.spring;
 import com.microservicio.fisibet.infraestructure.model.AccountUserModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountUserSpringPort extends JpaRepository<AccountUserModel, Long> {
+public interface AccountUserSpringPort extends JpaRepository<AccountUserModel, Integer> {
 }
